move sidebar and main positioning to layout; move main content to App.js
